+++ relearn/source/neurons/Neurons.cpp
@@ -463,15 +463,15 @@ std::vector<RankNeuronId> Neurons::delete_synapses_find_synapses_on_neuron(
 
     RelearnException::check(num_synapses_to_delete <= number_synapses, "Neurons::delete_synapses_find_synapses_on_neuron:: num_synapses_to_delete > current_synapses.size()");
 
-    std::vector<unsigned int> drawn_indices{};
+    std::vector<size_t> drawn_indices{};
     drawn_indices.reserve(num_synapses_to_delete);
 
     std::uniform_int_distribution<unsigned int> uid{};
 
     for (unsigned int i = 0; i < num_synapses_to_delete; i++) {
-        auto random_number = RandomHolder::get_random_uniform_integer(RandomHolderKey::Neurons, 0, number_synapses - 1);
+        auto random_number = RandomHolder::get_random_uniform_integer(RandomHolderKey::Neurons, size_t(0), number_synapses - 1);
         while (std::find(drawn_indices.begin(), drawn_indices.end(), random_number) != drawn_indices.end()) {
-            random_number = RandomHolder::get_random_uniform_integer(RandomHolderKey::Neurons, 0, number_synapses - 1);
+            random_number = RandomHolder::get_random_uniform_integer(RandomHolderKey::Neurons, size_t(0), number_synapses - 1);
         }
 
         drawn_indices.emplace_back(random_number);
